refactor: move UvicornServer & co to llama_cpp.py
The backends module, along with the llama_cpp module,
have a circular dependencies, which is considered an anti-pattern.

Move UvicornServer and get_uvicorn_config
to reduce the circular dependency
llama_cpp.py -> backends.py.
